src/backend/requirements.txt
fastapi
google-api-core
google-auth
google-cloud-bigquery
google-cloud-discoveryengine
httpx
pydantic
pytest
pytest-asyncio
pytest-cov
pyyaml
uvicorn
